fix(chat): Add emoji shortcodes to message reaction labels
