+++ bigbluebutton-html5/imports/ui/components/chat/chat-graphql/chat-message-list/page/chat-message/message-reactions/styles.ts
@@ -1,36 +1,4 @@
 import styled from 'styled-components';
-import {
-  colorGrayLighter, colorGrayLightest, colorOffWhite,
-} from '/imports/ui/stylesheets/styled-components/palette';
-
-const EmojiWrapper = styled.button<{ highlighted: boolean }>`
-  background: none;
-  border-radius: 1rem;
-  padding: 0.375rem 1rem;
-  line-height: 1;
-  display: flex;
-  flex-wrap: nowrap;
-  border: 1px solid ${colorGrayLightest};
-  cursor: pointer;
-
-  ${({ highlighted }) => highlighted && `
-    background-color: ${colorOffWhite};
-  `}
-
-  em-emoji {
-    [dir='ltr'] & {
-      margin-right: 0.25rem;
-    }
-
-    [dir='rtl'] & {
-      margin-left: 0.25rem;
-    }
-  }
-
-  &:hover {
-    border: 1px solid ${colorGrayLighter};
-  }
-`;
 
 const ReactionsWrapper = styled.div`
   display: flex;
@@ -40,6 +8,5 @@ const ReactionsWrapper = styled.div`
 `;
 
 export default {
-  EmojiWrapper,
   ReactionsWrapper,
 };
